+++ db_scripts/06_CrearProcedimiento_de_Modificar_de_Usuario-modificado.sql
@@ -48,8 +48,8 @@ END
 GO
 
 /* =======================
-   PRUEBAS (SSMS) - OPCIONALES
-   Ejecutar por bloques:
+   PRUEBAS (SSMS) - OPCIONALES (Descomentar para usar)
+   Ejecutamos por bloques seleccionando y presionando F5
    ======================= */
 -- 1) Ver antes
 -- SELECT * FROM dbo.Perfiles WHERE CodigoUsuario = 1000;